Migrated the generic execution service framework from the PipeOnline to Pipeline Core.
The main role of this execution service is to provide an abstract layer for executing and listen to Pipeline jobs, whatever the underlying model job object is, and which does not require dependencies on the Pipeline libs other than this execution package.

The framework contains a generic abstract execution service implementation based on the adapter pattern, and a subclass of this using a pool of RMI Pipeline instances to actually execute the jobs in parallel.