Fixed a bug in the SequenceSegment renaming algorithm:
Problem description:
When using the pattern "fixed(name)+seq" in a NCX-only DAISY 3 book, sequence numbers were not incremented and the strategy is resolved to:
name001
name001_
name001__
name001___

Problem cause:
The algorithm used when creating a SequenceSegment was:

1.	get the sequence number of a file in the fileset {
2.		get the position of the file in the fileset  { // line 63
3.			get an initial pool of files from the fileset manifest { // lines 110-120
4.				for each file in the initial pool {
5.					initialize caches of searched files (all files + files of the same type used as a counter) // lines 127-128
6.					search recursively the searched file in the current file and its referenced local members // lines 144-178
7.				}
8.			} 
9.		}
10.		return the formatted position
11.	}

 => line 5. of this algorithm should just be moved out of the loop.

The initial pool of searched files is the manifest spine (ie. the smil files). In the case of NCX-only books, each smil file only references an audio file, but not the dtbook. Hence the searched file is not found in the first recursive search. Because the caches are reinitialized before we search the next file in the original pool, the code always returned a sequence number equals to 1.