Tried to fix the long-standing "duplicate doctitle" issue:
The Pipeline no-longer systematically creates a frontmatter (was doing that in the tidy-add-author-title). The new behavior is:

b1. when the frontmatter is present but has no doctitle (possible in DTBook 2005-1 or invalid  2005-2 and 2005-3 DTBooks), a doctitle is added from the metadata value. I can possibly change it to pick the first doctitle or heading found in the bodymatter if it's a preferred solution.

b2. when the frontmatter is present but has no docauthor, a docauthor is inserted with the value of the metadata. It is a long-standing "tidy" step made by the Pipeline, which is worth keeping IMHO.

b3. if no frontmatter is present, it is kept as-is, no artificial frontmatter is added (regardless of whether the bodymatter includes a doctitle or not).

b4. if the "dc:Title" metadata is not set (invalid), the value is set to the value of the first doctitle, or the first heading if no doctitle is found.