Add some code to handle excercise markup as defined in the nordic requirements
Essentially there are a few classes that define how to mark up
excercises. The stuff that is relevant for large print is how to
present answer fields, i.e. placeholders in the original book to fill
in an answer or check a box in a multiple choice question. The tree
classes that are defined ("answer", "answer_1" and "box") are now
supported.