Large Print: Try to find a workable solution for inline prodnotes
The tcolorbox environment works fine in the block context but for
inline prodnotes it is not really an ideal solution, is it typesets
the prodnote in its own paragraph or if inline it makes it
non-breakable. A solution for inline prodnotes could be to treat them
like footnotes, but it doesn't seem very desirable to typeset block
and inline prodnotes differently. So for now inline prodnotes are set
in a separate paragraph like their block context brother. However they
are not floating, i.e. interrupt the current paragraph immediately.