Large Print: Do not produce automatic text for captions
i.e. we do not want the usual LaTeX style of "Table 4:...". Instead we
want just the content of the caption. For that we use the \legend
macro from the memoir class. This has the added benefit that it also
works outside of floats. Since all the text is left justified and we
would like to stay away from italics there is no visual clue that
something is a caption. It looks just like a normal paragraph.