Add support for span@class='linenum'
The span@class='linenum' is a "proprietary extension" of SBS if you will
that allows to markup books that have very long passages of numbered lines,
i.e. basically spanning the whole book. Linegroup and line is not suited
for this as you can no longer markup paragraphs, blockquotes or anything
else for that matter.