Handle headings inside frontmatter properly
When inside the frontmatter we do not want a heading to be listed in
the toc. That is why we use the starred version of the sectioning
commands, i.e. section*{foo}. However the starred versions of these
commands do not take optional arguments, so do not insert them when
inside the front matter.

This is a fixup for 388afac3c1bdd44289d26c1d58f385c49b7e9473.