Make the addition of hyphenation points more robust
before it would add hyphenation points to any long sequence of
characters. This can be problematic if the word contains characters
that need to be excaped for LaTeX. So a word like [foo] would be
escaped to \lbrack{}foo\rbrack{} and then soft hyphend would be added
to that construct. Which of cource fails miserably.

Now we scan the text for long words before escaping and only hyphenate
the word itself, excluding any punctuation before and after the word.